Add support for conditional nodes in cuda graphs.
This allows torch.cond and torch.while_loop to be captured in a single
cuda graph. This is done by manually inserting conditional IF nodes
and conditional WHILE nodes during stream capture. This approach is
discussed here:

https://developer.nvidia.com/blog/dynamic-control-flow-in-cuda-graphs-with-conditional-nodes

Previously, data-depenent control flow would force the
usage of cuda graph trees, since data-dependent control flow was done
on the CPU. Now, data-dependent control flow is done on the GPU.

This work depends upon CUDA 12.4, since cuda graph conditional nodes
were introduced in CUDA 12.4.

This works only with torch.compile(..., backend="eager") and
torch.compile(..., backend="cudagraphs") backends right now. Notably,
there is no inductor support at this time!

Conditional nodes for cuda graphs were first experimented with in
https://arxiv.org/abs/2406.03791 . While this paper showed strong
improvements in data-dependent workloads that were very CPU-overhead
bound, the next place to look for improvement is horizontal and
vertical kernel fusion, which can eventually be enabled automatically
once conditional nodes work with backends like inductor. This PR is
the first step towards that. I also expect this work to benefit more
sophisticated models like autoregressive decoding of LLMs, at least
for users that are using static shape kv-caches.

This is work done by @tingyangk and me.

We have a sophisticated example of RNN-T greedy decoding (the
algorithm discussing in the paper) working with this new feature here:
https://github.com/tingyangk/NeMo/commit/975a80673edff84a78097bb370fe2d37316e5ad6#diff-2c2a72c9a5392d4a6ea5149fea3ce7900b9fd2c630e460bbab94547379553ceaR376

Some design decisions:

1. We initially implemented this by using the pypi cuda-python
package. We changed to modifying CUDAGraph.cpp instead, and wrapping
the new methods in pybind11 to expose them in python. This would
presumably allow for triton's C++ code generator to call the relevant
APIs. It also avoids a new dependency on cuda-python. As a nuisance,
the "bodies" of conditional nodes need to be captured to new streams.
Therefore, we need access to the stream capture mode of the currently
capturing stream to pass to cudaStreamBeginCaptureToGraph().

2. We expose the currently capturing CUDAGraph via a thread_local
variable with the static method
CUDAGraph::get_currently_capturing_graph(). Basically, since we don't
pass the CUDAGraph at every step of a particular computation, we need
to be able to access it to call the newly exposed APIs. There is a
long doc string in CUDAGraph.cpp about this, but the TL;DR is that it
may be more correct to have a global hash table mapping CUDAStreams to
CUDAGraphs instead. It depends on whether or not pytorch code is ever
expected to stream capture to a graph from more than one thread. I
don't believe that any pytorch code today does this, but I may be
mistaken.